feat!: Add enforcing entity association loader
